test_xai_chat_config_get_openai_compatible_provider_info
